+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ### vNext
 - Change package to produce ES2015 as `module` and commonjs for `main` [#1576](https://github.com/apollographql/react-apollo/pull/1576)
+- Make `Query` component work with `getDataFromTree` by defining `fetchData` [#1579]
 
 ### 2.1.0-alpha.0