+++ Changelog.md
@@ -1,9 +1,12 @@
 # Change log
 
 ### vNext
-- Change package to produce ES2015 as `module` and commonjs for `main` [#1576](https://github.com/apollographql/react-apollo/pull/1576)
-- Make `Query` component work with `getDataFromTree` by defining `fetchData` [#1579]
-- Added back in support for browser / main bundles [#1578](https://github.com/apollographql/react-apollo/pull/1578)
+
+### 2.1.0-alpha.1
+
+* Change package to produce ES2015 as `module` and commonjs for `main` [#1576](https://github.com/apollographql/react-apollo/pull/1576)
+* Make `Query` component work with `getDataFromTree` by defining `fetchData` [#1579]
+* Added back in support for browser / main bundles [#1578](https://github.com/apollographql/react-apollo/pull/1578)
 
 ### 2.1.0-alpha.0